+++ example/lib/your_model.dart
@@ -6,26 +6,28 @@ part 'your_model.g.dart';
   Property<String?>('nullableString', value: '"123"'),
   Property<int>('fixInt'),
   Property('withValueConvert', value: '12'),
-  Property<List<String?>?>('listWithType'),
+  Property<List<String>?>('listWithType'),
   Property<List?>('listNoType'),
   Property<Map<String?, dynamic>?>('mapWithType'),
   Property<Map?>('mapNoType'),
+  Property<DateTime>('dateTime'),
 ])
 class YourModel extends _YourModelImpl {
-
   YourModel([super.data]);
 }
-
 
 @Binding([
   Property<String?>('nullableString', value: '"123"'),
   Property<int>('fixInt'),
   Property('withValueConvert', value: '12'),
-  Property<List<String?>?>('listWithType'),
+  Property<List<String>?>('listWithType'),
   Property<List?>('listNoType'),
   Property<Map<String?, dynamic>?>('mapWithType'),
   Property<Map?>('mapNoType'),
-])
+  Property<DateTime>('dateTime'),
+], converts: {
+  Map<String?, dynamic>: 'castMap',
+})
 class SuperBinding extends _SuperBindingImpl {
   SuperBinding([super.data]);
 }
@@ -35,4 +37,11 @@ class SuperBinding extends _SuperBindingImpl {
 ])
 class SubBinding extends SuperBinding with _SubBindingMixin {
   SubBinding([super.data]);
+}
+
+Map<String?, dynamic> castMap(String property, dynamic value) {
+  if (property == 'mapWithType') {
+    // hit Field
+  }
+  return value;
 }