+++ model_binding/lib/src/widget/text_field_binding.dart
@@ -143,9 +143,9 @@ class TextFieldBindingState extends State<TextFieldBinding> {
     if (value == null) return;
     var newValue = value;
     if (widget.convert != null) newValue = widget.convert!(value);
-    var oldValue = widget.binding.$_data[widget.property];
+    var oldValue = widget.binding.$data[widget.property];
     if (newValue != oldValue) {
-      widget.binding.$_data[widget.property] = value;
+      widget.binding.$data[widget.property] = value;
       if (widget.onChanged != null) widget.onChanged!(value);
 
       var widgetContext = widget.context;
@@ -230,7 +230,7 @@ class TextFieldBindingState extends State<TextFieldBinding> {
         spellCheckConfiguration: widget.spellCheckConfiguration,
         magnifierConfiguration: widget.magnifierConfiguration,
         controller: widget.binding.textField(widget.property,
-            value: widget.binding.$_data[widget.property],
+            value: widget.binding.$data[widget.property],
             retainSelection: true,
             convert: widget.convert),
       );