+++ model_binding_builder/CHANGELOG.md
@@ -21,3 +21,10 @@
 
 ## 1.0.7
 - No longer providing convert, but verifying the type.
+
+## 1.0.8
+- Add default type convert.
+- Add custom type convert.
+
+## 1.0.9
+- update comments.